services/api/pyproject.toml
[project]
name = "eha-api"
version = "1.0.0"
description = "EHA - Email Helper Agent API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "celery[redis]>=5.3.6",
    "redis>=5.0.0",
    "structlog>=24.1.0",
    "google-api-python-client>=2.115.0",
    "google-auth>=2.27.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "jsonschema>=4.21.0",
    "PyNaCl>=1.5.0",
    "firebase-admin>=6.4.0",
    "apns2>=0.7.2",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
